:sunrise: Initial commit
The first commit with a fully-configured Hugo site, ready to host
content as we described earlier today in the Design & UX Working Group
meeting for December 2021. This uses the UNICEF Inventory theme for the
site look-and-feel.

The initial categories are created but do not hold any content yet. Some
of the configuration values may change in the future.

Signed-off-by: Justin W. Flory (he/him) [UNICEF Innovation] <jflory@unicef.org>