:new: :memo: guidelines: Add reading list from UNICEF Inventory (#3)
Closes unicef/inventory#83.

This commit imports the "Design & UX Reading List" from the UNICEF Open
Source Inventory. This moves the slim range of content on design from
the Inventory and into the SustainOSS Design portal. It also provides an
example of what a new article or post might look like.

Once this is added to the SustainOSS Design portal, the page and
category should be removed from the UNICEF Inventory to avoid
duplication of content.

Signed-off-by: Justin W. Flory (he/him) [UNICEF Innovation] <jflory@unicef.org>