Update _index.en.md
It is quite subjective and imho rather offensive to prioritize as "most notably," the difference between design that serves aesthetic needs and design that serves interactive needs. Likewise, a not insignificant bulk of information was absent, that is important—and use of the word "pretty" when we are defining who designers are and aren't, is quite loaded and imho needlessly dismissive. 

This may be overkill, but submitting, anyway. Historically, OSS has been more exclusionary than inclusive; and I'd like to see that change.