:new: :two_hearts: .github: Opt in to SustainOSS Code of Conduct
This commit adds a Code of Conduct file to the repository, which then
points to the SustainOSS Code of Conduct on `sustainoss.org`. This our
explicit opt-in to the same CoC as the wider SustainOSS community.

Since we are already getting exposure and interest from folks beyond our
core Working Group team, I believe it is important we put this in place
now to explicitly state our values as a team and project, and to give
some intermediary advice on what to do if there are issues.

Signed-off-by: Justin W. Flory (he/him) [UNICEF Innovation] <jflory@unicef.org>