:new: :busts_in_silhouette: .github: Add contributing guidelines
This commit adds contributing guidelines that explains three topics:

1. Contribution process
2. Structure and components
3. How to create a development environment

The contribution process suggests we use a governance model known as
**lazy consensus**. This model allows a change or new idea to move
forward, provided it gets at least one "+1" vote and no "-1" vote from a
maintainer. The full context is explained in the document.

Additionally, this document also explains how to create a development
environment and set up the project. It explains how to clone the git
repository and set up the git submodule for the theme, and how to run
the Hugo server to get a local preview.

This addresses the feedback from @RichardLitt in PR #11 and should
unblock that Pull Request as well.

Signed-off-by: Justin W. Flory (he/him) [UNICEF Innovation] <jflory@unicef.org>