:bug: theme: Match navbar color to site primary color
Previously, the navbar at the top of the site was hard-coded to use the
UNICEF Blue color on all pages except the home page. This commit
introduces a fix added to the upstream theme to match the navbar color
with the site primary color, as set in the config file.

It also incorporates some improvements to CSS and the PDF download
option worked on by @Neha9849.

ref: unicef/inventory-hugo-theme#58

Signed-off-by: Justin W. Flory (he/him) [UNICEF Innovation] <jflory@unicef.org>