:wrench: :link: config: Add menu link for SOS Design podcast
Related to #20 but does not close it.

This adds a new menu to the site. This causes the SOS Design podcast
link to show up at the top of the page but also in the footer too. It is
the easiest way to get more visibility for another site or URL.

Signed-off-by: Justin W. Flory (he/him) [UNICEF Innovation] <jflory@unicef.org>